SYZ Signpost fix
The third spotlight 'face' is unused, and is the only one that is used for the signpost. It's been commented out.

The signpost resets the face/vertex count cause the spotlight also resets it when drawing itself.

The spotlights need to stop drawing when close to the signpost, as two 3DScene draw calls messes both of them up.